ajustando o codigo

+++ src/main/java/org/example/dao/ProdutoDao.java
@@ -3,6 +3,7 @@ package org.example.dao;
 import org.example.modelo.Produto;
 
 import javax.persistence.EntityManager;
+import java.math.BigDecimal;
 import java.util.List;
 
 public class ProdutoDao {
@@ -48,11 +49,9 @@ public class ProdutoDao {
                 .getResultList();
     }
 
-    public Integer buscarPrecoDoProdutoComNome(String nome) {
+    public BigDecimal buscarPrecoDoProdutoComNome(String nome) {
         String jpql = "SELECT p.preco FROM Produto p WHERE p.nome = :nome";
-        
-        
-        return em.createQuery(jpql, Integer.class)
+        return em.createQuery(jpql, BigDecimal.class)
                 .setParameter("nome", nome)
                 .getSingleResult();
     }
